+++ src/Mapbender/CoreBundle/Element/FeatureInfo.php
@@ -75,6 +75,8 @@ class FeatureInfo extends AbstractElementService
             'strokeColorHover' => 'rgba(255,0,0,0.7)',
             'strokeWidthDefault' => 1,
             'strokeWidthHover' => 3,
+            'pointRadiusDefault' => 7,
+            'pointRadiusHover' => 9,
             'fontColorDefault' => '#000000',
             'fontColorHover' => '#000000',
             'fontSizeDefault' => 12,